requirements.txt
alembic==1.4.3
appdirs==1.4.4
asyncpg==0.21.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.6.20
chardet==3.0.4
cleo==0.8.1
click==7.1.2
clikit==0.6.2
crashtest==0.3.1
databases==0.4.0
distlib==0.3.1
filelock==3.0.12
h11==0.11.0
html5lib==1.1
idna==2.10
keyring==21.4.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
packaging==20.4
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.6.0
poetry==1.1.4
poetry-core==1.0.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pylev==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
requests-toolbelt==0.9.1
shellingham==1.3.2
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.8
tomlkit==0.7.0
urllib3==1.25.11
uvicorn==0.12.2
virtualenv==20.1.0
webencodings==0.5.1
